FIXME: Allow "invalid" weak unrefs when processing weak notifies
Epiphany is crashing because both AdwSpinnerPaintable and AdwAnimation
have taken weak refs on the AdwSpinner. When the AdwSpinner is
destroyed, the AdwSpinnerPaintable's weak notify executes first, and
destroys the AdwAnimation. The AdwAnimation, being disposed, then
attempts to unregister its own weak ref on the AdwSpinner, which fails
because the weak references have already been cleared from the datalist.
But AdwAnimation cannot really do anything about this, because its weak
notify has not yet been called. (It's about to be called, just as soon
as

--- wait, do not commit. Because it's about to be called. But the
AdwAnimation has already been destroyed. There is a use after free here,
and this doesn't fix it.

Fixes #3434