gdataset: reorder epilogue code path in g_datalist_id_update_atomic()
The most common case is that the callback does not change the
data/destroy pointers. Reorder the "if" checks to favor those cases.

g_datalist_id_update_atomic() makes sense to track some data structure,
and atomically update it. But in that case, we usually update the values
inside the structure, and don't reset the pointer to the structure
itself. That means, if old data was present, then new_data is likley
unchanged.

With this change, we instead check whether the values are unchanged and
avoid resetting the (same) value. The idea is in the common case to
avoid writing the same value to memory to avoid false sharing or
invalidating CPU caches.