Revert "CI: Mark msys2-mingw32 as allowing failures"
This reverts commit 16819a40245a7d844382cab55ee4725f33718bab.

The failure this was added for is intermittent and has been going on a
long time; it’s not a new failure caused by msys2 dependency/packaging
changes, so the policy in .gitlab-ci.yml doesn’t apply.

It would be great if someone fixed #3042, but un-gating GLib from msys2
testing is not the right tradeoff for it.

See: #3042