Revert "gcancellable: Mark assert-only variable as unused"
This reverts commit 6c679fb37a0bfeed7beb42d1b1e3590713eac10c.

Merge request !2765 has caused a thread safety regression in
`GCancellableSource` disposal. It landed too late in the cycle to fix
with any confidence before the 2.82 release, so the changes are being
reverted for the `glib-2-82` branch and 2.82.0 release.

This commit is not problematic, but is built entirely on top of the
problematic MR, so has to be reverted.

Fixes: #3448
Re-opens: #774, #2309, #2313