gio/gdbusmessage: Avoid most calls to _get_type()
This uses the type_id registered in the previous commit to avoid many
calls to g_dbus_message_get_type() for type checking within the source
file. We can rely on the type being registered once an object is created
and if not, there is no way the type check could succeed anyway.