Collation keys are not encoded in UTF-8
The value returned when generating a collation key is an opaque binary
blob that is only meant to be used for byte-wise comparisons; we should
not imply it's a nul-terminated, UTF-8 string.

This is especially true for language bindings that try to convert C
strings returned by GLib into UTF-8 encoded strings.

Ideally, the collation functions should return a byte array, but the
closest thing we have is the OS native encoding type that we use for
paths and environment variables.

See: https://github.com/gtk-rs/gtk-rs-core/issues/1504